requirements.txt
aiohttp==3.8.4
fastapi==0.95.2
httpx==0.24.1
hypercorn==0.14.3
pydantic==1.10.8
pytest==7.3.1
python-dotenv==1.0.0
